Fix TreeView vertical alignment issue
Previously, if certain children were expandable and others were not,
then they would not start have the same padding even though they were
at the same level. Now, if any items that share the same parent are
expandable then it will use the same amount of padding for each of them,
even for the nodes that are not expandable.

Modify example to add a case that reproduces this.