DynamicOverflow widget
Similar to Wrap, but only displays widgets on a single run, and then
hides widgets that would have overflown. If widgets are hidden, then it
will display the specified "overflow widget".

This is useful to implement the dynamic overflow mode of command bars.

It could also be useful if a vertical nav bar wanted to implement a "more"
button instead of vertical scrolling.